# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /architectures/{id}/versions:
    get:
      operationId: listVersions
      tags: [versions]
      summary: List version history for an architecture (newest first)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArchitectureVersionSummary"
  /architectures/{id}/versions/{versionId}:
    get:
      operationId: getVersion
      tags: [versions]
      summary: Get one version including its full snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Version with snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchitectureVersion"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /architectures/{id}/versions/{versionId}/rollback:
    post:
      operationId: rollbackToVersion
      tags: [versions]
      summary: Restore the architecture to a previous version's snapshot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: versionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Architecture after rollback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Architecture"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /validations/analyze:
    post:
      operationId: validateInfrastructure
      tags: [validations]
      summary: Analyze infrastructure code for security, reliability, cost, scalability, and operational issues (SSE stream)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateInfrastructureInput"
      responses:
        "200":
          description: SSE stream of validation analysis
          content:
            text/event-stream: {}
...
components:
  schemas:
...
    CloudProvider:
      type: string
      description: Target cloud provider
      enum: [aws, azure, gcp]
...
    ArchitectureVersionSummary:
      type: object
      required: [id, architectureId, versionNumber, reason, changedFields, editorName, createdAt]
      properties:
        id:
          type: integer
        architectureId:
          type: integer
        versionNumber:
          type: integer
        reason:
          type: string
        changedFields:
          type: array
          items:
            type: string
        editorName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    ArchitectureVersion:
      allOf:
        - $ref: "#/components/schemas/ArchitectureVersionSummary"
        - type: object
          required: [snapshot]
          properties:
            snapshot:
              type: object
              additionalProperties:
                type: string
              description: Content fields as they were before the change
...
    Architecture:
      type: object
      required: [id, title, requirements, provider, diagram, terraform, costEstimate, securityRecommendations, highAvailabilityPlan, databaseRecommendation, kubernetesDeployment, cicdPipeline, monitoringSetup, disasterRecovery, threatModel, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        requirements:
          type: string
        provider:
          $ref: "#/components/schemas/CloudProvider"
        diagram:
          type: string
          description: Mermaid diagram source
        terraform:
          type: string
          description: Terraform HCL content
        costEstimate:
          type: string
        securityRecommendations:
          type: string
        highAvailabilityPlan:
          type: string
        databaseRecommendation:
          type: string
        kubernetesDeployment:
          type: string
        cicdPipeline:
          type: string
        monitoringSetup:
          type: string
        disasterRecovery:
          type: string
        threatModel:
          type: string
          description: STRIDE threat model as a JSON document
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ValidateInfrastructureInput:
      type: object
      required: [source, format]
      properties:
        source:
          type: string
          description: Raw infrastructure code to analyze (uploaded or pasted)
        format:
          type: string
          description: Infrastructure code format
          enum: [terraform, cloudformation, kubernetes]
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string