# lib/api-zod/src/generated/types/architectureVersion.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArchitectureVersionSnapshot } from './architectureVersionSnapshot';
import type { ArchitectureVersionSummary } from './architectureVersionSummary';

export type ArchitectureVersion = ArchitectureVersionSummary & {
  /** Content fields as they were before the change */
  snapshot: ArchitectureVersionSnapshot;
};