# lib/api-zod/src/generated/types/architectureVersionSummary.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArchitectureVersionSummary {
  id: number;
  architectureId: number;
  versionNumber: number;
  reason: string;
  changedFields: string[];
  /** @nullable */
  editorName: string | null;
  createdAt: Date;
}